+++ packages/address/src/AddressFormatter.ts
@@ -1,4 +1,10 @@
-import {Address, FieldName, Country} from './types';
+import {
+  Address,
+  FieldName,
+  Country,
+  SupportedLocale,
+  SupportedCountry,
+} from './types';
 import {renderLineTemplate, FIELDS_MAPPING} from './utilities';
 import {loadCountry, loadCountries} from './loader';
 
@@ -20,17 +26,17 @@ const COUNTRIES_CACHE: {
 } = {};
 
 export default class AddressFormatter {
-  constructor(private locale: string) {
-    this.locale = locale.toUpperCase();
+  constructor(private locale: SupportedLocale) {
+    this.locale = locale;
     COUNTRIES_CACHE[this.locale] = {};
   }
 
-  updateLocale(locale: string) {
-    this.locale = locale.toUpperCase();
+  updateLocale(locale: SupportedLocale) {
+    this.locale = locale;
     COUNTRIES_CACHE[this.locale] = {};
   }
 
-  async getCountry(countryCode: string): Promise<Country> {
+  async getCountry(countryCode: SupportedCountry): Promise<Country> {
     let country = this.loadCountryFromCache(countryCode);
     if (country) {
       return country;
@@ -38,6 +44,7 @@ export default class AddressFormatter {
 
     country = await loadCountry(this.locale, countryCode);
     COUNTRIES_CACHE[this.locale][country.code] = country;
+
     return country;
   }
 
@@ -83,7 +90,9 @@ export default class AddressFormatter {
   *     ['phone']
   *   ]
   */
-  async getOrderedFields(countryCode: string): Promise<FieldName[][]> {
+  async getOrderedFields(
+    countryCode: SupportedCountry,
+  ): Promise<FieldName[][]> {
     const country = await this.getCountry(countryCode);
 
     const format = country ? country.formatting.edit : DEFAULT_FORM_LAYOUT;
@@ -100,7 +109,7 @@ export default class AddressFormatter {
   }
 
   async getTranslationKey(
-    countryCode: string,
+    countryCode: SupportedCountry,
     key: FieldName,
   ): Promise<string> {
     const country = await this.getCountry(countryCode);
@@ -118,7 +127,7 @@ export default class AddressFormatter {
   }
 
   private loadCountryFromCache(
-    countryCode: string,
+    countryCode: SupportedCountry,
   ): Country | undefined | null {
     const cachedCountry = COUNTRIES_CACHE[this.locale][countryCode];