+++ packages/address/CHANGELOG.md
@@ -17,3 +17,6 @@ The format is based on [these versioning and changelog guidelines](https://git.i
 - Changed `address2Key: string` type to `ZipKey` type ([#274](https://github.com/Shopify/quilt/pull/274))
 - Removed `attributes` key from Country type ([#274](https://github.com/Shopify/quilt/pull/274))
 - Changed `format` key to `formatting` in Country type ([#274](https://github.com/Shopify/quilt/pull/274))
+- Created `SupportedLocale` type ([#311](https://github.com/Shopify/quilt/pull/311))
+- Created `SupportedCountry` type ([#311](https://github.com/Shopify/quilt/pull/311))
+- `AddressFormatter` now accepts `SupportedLocale` as locale instead of `string` ([#311](https://github.com/Shopify/quilt/pull/311))